Bug #22: Fix confusion involving Endpoint: field in General tab of SsgPropertyDialog. An Ssg's localEndpoint property is now filled in on-the-fly as ssgNNN as needed, where NNN is the Ssg's ID.
Also the dialog box will put a leading slash in for you, but the Ssg bean will strip it off
again if it's there.  This should eliminate that source of trouble, too.

At the moment the localEndpoint property is read-only, but it can be switched back to read/write
by changing a JLabel back into a JTextField (a two-gesture job in IDEA).


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@1165 219f1c6e-4a0e-43c8-96f2-6e749be13fd4