We now maintain one KeyStore on disk per SSG.  We can quickly determine if an SSG has a client certificate, and fairly quickly access it's certificate, public key, and private key (using the SSG password).
Need to add code to generate the key pair, then hook up the CSR code,
then post the CSR to the SSG and download the resulting cert.  Then,
need to brain up the KeyManager to present the appropriate client
certificate during the SSL handshake.  Might use a thread-local
variable for this; other alternative is a hash mapping SSG hostnames to corresponding SSG objects.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@1703 219f1c6e-4a0e-43c8-96f2-6e749be13fd4