Found a gross hack that helps alleviate the problem with finding the logon dialog when it pops up underneath all the other windows.
On Windows XP this will work if the app is merely in the background, but not if it's actually been minimized -- XP appears to prevent minimized programs from popping themselves back up and into the foreground.  In this case it will at least flash the task bar so the user has some hope of noticing it.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@2339 219f1c6e-4a0e-43c8-96f2-6e749be13fd4