First version of SecureSpanAgent and SecureSpanAgentFactory, the documented external interfaces for using the Agent as a Java class library.
Moved the server and client cert utility methods from ClientProxy into SsgKeyStoreManager.

The obtainClientCert() procedure now requires the CSR to be regenerated for each retry in the case of a bad password.  This simplifies flow, fixes a bug
involving the username changing but the CSR not getting updated, and allows
the method to be called in contexts where there is not yet any pending request.

The PolicyManagerImpl now logs policies it is using if the system property
com.l7tech.proxy.datamodel.logPolicies is set to "true".

PendingRequest now has a prepareClientCert() method that ensures that the
client cert is available.  Intended to be used by the client assertions.

SsgResponse now remembers what HTTP status was returned by the Gateway.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@4016 219f1c6e-4a0e-43c8-96f2-6e749be13fd4