Further work on TROGDOR integration.
The big-but-not-very-useful asXmlString() methods were removed from the WssProcessor.  If the caller really wants a parsed element in string form they can call XmlUtil.nodeToString() for themselves.

X509 security tokens now keep track of whether a proof-of-possesion was seen in the form of at least one successfully-validated signature, as well as a possibly-unneeded list keeping track of which elements were signed with this token.  In turn, elements that were signed now keep track of which security token had been used to sign them.

The new TROGDOR-dependent ReceiverXmlSecurityProcessor is complete but untested.

TROGDOR now keeps a hashmap of x509 tokens it has read, indexed by wsu:Id.  Fixed NPE when a Signature referred to an as-yet-undeclared BinarySecurityToken.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@6711 219f1c6e-4a0e-43c8-96f2-6e749be13fd4