We now save copies of the encrypte elements before letting Marth mangle them so we can compare them with what comes out the other side.  There is still evidently a bug somewhere, probably in xss4j, since the "productid" element in the example message is causing the xss4j exclusive canonicalizer to NPE.  A stand-alone repro for this apparent xss4j bug is in CanonicalizerBugRepro.
There is still a problem with us failing to remove wsu:Id and xmlns:wsu attributes that we added to user elements during decoration.  While this hopefully won't matter in actual operation (WSSE has a catch-all CYA clause to the effect that "any element whose schema forbids adding a wsu:Id global attribute is just not allowed to be signed directly") it is causing the exact comparison in the SkilessRecipientCert test case to fail.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@6726 219f1c6e-4a0e-43c8-96f2-6e749be13fd4