Further work on policy encapsulation.
Added addChild() and removeChild() to CompositeAssertion, and started auditing usages of getChildren() to see who is going to break when the returned List is made into a read-only Collection.

It looks like a special iterator will be required in order to allow modification through it while preserving the policy tree invariants (parent pointers and ordinals).


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@7998 219f1c6e-4a0e-43c8-96f2-6e749be13fd4