Changed SslUtils to use SimpleHttpClient for password change action (isntead of making its own client lacking the proper SSL socket factory).
Renamed the Gui.errorMessage() method that suggests that shutting down the app might be necessary, and includes a "Shutdown" button on its dialog, to Gui.criticalErrorMessage() and added javadoc steering developers to use one of the other error reporting routines if the error isn't so critical after all.

Fixed server certificate disco when it is triggered by a password change action.  Other things in the Bridge that directly or indirectly cause SSL connections to the Gateway should be checked to ensure that they still trigger server cert disco correctly if the server cert isn't yet known when the particular function is triggered.  (I'm worried about the WSDL proxy feature in particular, since it is basically never used by any developers and so is particularly prone to bitrot.)

Changed some error message reporting to avoid reporting as a "critical error, restart the agent" for conditions where restarting the agent will not help in any way (including but not limited to failed SSL connections, bad original password, bad new password, etc.)

After all this, I was finally able to attempt to reproduce Bug #1344, only to find that for whatever reason I can't now reproduce the behaviour described in the bug.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@10299 219f1c6e-4a0e-43c8-96f2-6e749be13fd4