Started skunkworking the FIM certificate workshop.  This is turning out to require no production code changes at all so far, other than the most trivial of tweaks to make the TokenServiceClient's namespace URIs configurable.
So far the token service client can repro the desired request and parse the example response, and trogdor can deal with the example client request.  The token service client isn't prepared to accept a username token in the response, though, and I see the ServerWsTrustCredentialExchange assertion doesn't use it for that, instead doing its own thing.  I haven't attempted to repro the server RST message.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@10694 219f1c6e-4a0e-43c8-96f2-6e749be13fd4