Created friendly hand-holding GUI for managing licenses from the SecureSpan Manager. It's currently in the Help menu because that seemed most appropriate.
Added two more methods to ClusterStatusAdmin to query and install licenses.  These have a hotline direct to the license manager so that the GUI can find out if the license manager has a beef with the currently installed license and, if so, what its problem is.  (Naturally, installing a new license requires admin access to the cluster.)

Added DateUtils for common date-related routines.

Split InvalidLicenseException out of LicenseException.  The former is now for invalid 
licenses that couldn't be installed, while the latter is only for reporting operations that are not permitted by the currently installed valid license.

License now encapsulates more of the features of the license file, so it'll be possible to enhance the license format to support things like multi-product/version licenses in the future without changing any other classes.

BuildInfo now contains a strictly formatted major and minor version number, maintained by the development manager by hand whenever a new branch is created.

DEVLICENSE.xml has been updated -- the license manager now enforces product names and versions.  It gets both of these from BuildInfo.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@11827 219f1c6e-4a0e-43c8-96f2-6e749be13fd4