Moved TextEntryPanel, UrlPanel, and ValidatedPanel to common.    This required moving Actions.setEsc.. and setEnter... to from Actions (console) to Utilities (common).
Added a "simplify" static method to Assertion.  After it is sufficiently well-tested (it has barely been tested against multi-branch policies, currently) it could be made available as a function in the SSM.  Simplifying a policy tree converts policies like this: AND(AND(AND(foo bar) baz) AND(blat OR(blee XOR() blah) AND() OR(bloo) XOR(bleet AND(bloat)) zee)) into this instead (which should be equivalent): AND(foo bar baz blat OR(blee blah) bloo XOR(bleet bloat) zee)

The Bridge now supports a new SSG type, "Generic web service".  This takes a POST URL instead of a Gateway hostname.  If you want it to do anything useful you need to import some policies manually.  To this end, the Bridge also supports importing policies from a WSDL: you type the WSDL URL, click on the binding you want, and the policy table gets populated with policies for each operation imported from the WS-SecurityPolicy in the WSDL.

Still TODO for "Generic web service" feature:
- don't offer proxied WSDL lookups since there aren't any
- add "import server cert" button
- disable server cert disco
- disable policy disco
- disable CSR client
- make the BRA benefit from the above features
- have a nicer diagram drawn up


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@13209 219f1c6e-4a0e-43c8-96f2-6e749be13fd4