bugz 3843: allows the system to be rebooted from ssgconfig menu (requires root password)
Also, java home is now determined interactively during package install and set by symlinking /ssg/jdk to the location found. All startup scripts, and anything that sets JAVA_HOME on solaris, now sets it to /ssg/jdk. This means we are no longer tied to a particular minor version of the JDK, in a particular, hardcoded location.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@16618 219f1c6e-4a0e-43c8-96f2-6e749be13fd4