Gateway can now run as a standalone jar, starting tomcat in embedded mode:
   java -Dcom.l7tech.server.home=c:/ssg -jar Gateway.jar

This requires that a directory /ssg/etc/inf/WEB-INF exists, but it need contain only web.xml and the Spring config files.

The existing war file and RPM have not (yet) been changed in any way and should still work fine (as long as the presence of the tomcat 6 jars in WEB-INF/lib won't cause trouble).

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@16825 219f1c6e-4a0e-43c8-96f2-6e749be13fd4