backup of cluster shared key in a separate process is more working now.
creation of the keystore when there is already an existing keystore works now ... with a few caveats. It seems that the db keystore isn't preserved when migrating from software keystore to HSM.

it also seems that, on occasion, the SCA locks up when migrating from HSM to HSM.
there's a bug logged to track this (4026)
more investigation needed.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@16951 219f1c6e-4a0e-43c8-96f2-6e749be13fd4