bugz 4137:
removal of the GUI configurator. Now, when launching the configwizard, the default is to launch the console. You have to specify the -graphical option to launch the GUI (which will no longer be maintained). I am not 100% removing it yet, because I remain confident it will be requested again within the year.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@17207 219f1c6e-4a0e-43c8-96f2-6e749be13fd4