Band-aid hack temporary fix for AutoTest failure privatekey.Bra testKeyEncryptReq.
The new WssProcessor iterates the parts while retaining a reference to a DOM.  This is a problem because any mutation of the message via MimeKnob invalidates any DOM or cursors previously obtained from XmlKnob.  The XmlKnob contract incorrectly states that this invalidation only occurs when you actually modify some part's bytes -- however, currently this actually happens as soon as anyone calls getParts(), getFirstPart(), or other MimeKnob methods.

The effect of having the DOM invalidated is to prematurely commit changes made to the writable DOM, and to disregard any further changes made to that DOM instance.  This probably broke many things, but specifically it caused later code in the Bridge that attempted to remove the security header to become ineffective, allowing fully decorated responses to leak back through to the client, causing the AutoTest failure.

Parenthetically, this extra DOM serialization is likely also the cause of the 10% performance hit observed earlier.

For now I have simply hacked the WssProcessor to forcibly restore the DOM after every use of MimeKnob methods.  A proper fix would be to have the MimeKnob wrappers in XmlKnob honor the current XmlKnob contract and commit DOM changes only when someone actually requests to access (or alter) the actual bytes of the first part.

A unit test for this particular problem was added to WssProcessorTest.  As well, some additional coverage of DOM invalidation was added to KnobblyMessageTest.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@17656 219f1c6e-4a0e-43c8-96f2-6e749be13fd4