(Bug #4524)
Upgraded JUnit to JUnit 4.4, and upgraded ant to 1.7.0 so its testrunner supports the new annotations.  

I found this article to be a great overview of useful features added in JUnit 4:

  http://today.java.net/pub/a/today/2006/12/07/junit-reloaded.html

Removed or replaced test code that relied on ancient versions of axis.jar and jaxrpc.jar and removed both jar files from lib/test.

Added a unit test for Bug #4542, annotated with @Ignore until the bug is fixed.

Fixed Bug #4524, and added some test coverage of the Bridge WSDL proxy component to prevent future regressions of this kind.

Introduced  an annotation @BugNumber we can use to link unit tests to a particular Bugzilla number in a machine-readable way.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@17934 219f1c6e-4a0e-43c8-96f2-6e749be13fd4