bug 4464:
shared key backup now works properly for both supported platforms. It can take a while on solaris (don't know why it's so slow) but it does finish and does work.

had to make some changes to the various scripts that use the ssg-utilities command so that they only ask for passwords and su when absolutely needed (i.e. not root or ssgconfig).

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@18072 219f1c6e-4a0e-43c8-96f2-6e749be13fd4