Bug 4667: The MemoizedRsaSha1SignatureMethod now includes the SignatureValue in the cache lookup key along with the signature hash and the public key bytes.
ServerConfig now prepopulates all system properties for serverconfig properties with ".setSystemProperty=true" as soon as the cluster property cache is attached to it.  

There is a new cluster property "rsasigcache.maxEntries" that can be used to size or disable the RSA signature cache; setting it to 0 prevents instances of MemoizedRsaSha1SignatureMethod from ever being created.  You must reboot the Gateway for changes to this value to take effect.

The signature cache size default to 2000.  We could consider changing the default value to zero to disable the cache by default.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@18168 219f1c6e-4a0e-43c8-96f2-6e749be13fd4