Bug 4970: The Bridge now issues a proxied WSDL/WSIL GET request without credentials if the user enters an empty username.
The Gateway's WSDL proxy servlet now serves up the anonymously-accessible services if a request arrives without credentials, as it used to do whenever authentication failed before the fix for Bug #4338.

If credentials are supplied with the request but do not authenticate or authorize, the Gateway will continue to return 403, for which the Bridge will continue to display the "Bad password" dialog.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@18608 219f1c6e-4a0e-43c8-96f2-6e749be13fd4