bug #4672
Removed the private AuthenticationResult constructor and the static AUTHENTICATED_UNKNOWN_USER variable that was referencing it.  This variable was only being used in the setAuthenticated method in ServerCustomAssertionHolder's CustomServiceRequest. 

Now when setAuthenticated is called, it just sets a boolean which is checked after the customServiceRequest's serviceInvocation has returned.  If this variable is true and principal credentials exist, then a new AuthenticationResult is created with that information.  

This not only fixed bug #4672, but also another bug where 2 authenticated users were appearing in the audit log when a custom assertion was being used for access control.  

This removes the need for an AuthenticationResult that contains an unknown user.

I updated the SecurityContext javadoc (which appears as part of the Custom Assertion API javadoc) to reflect the above changes.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@18690 219f1c6e-4a0e-43c8-96f2-6e749be13fd4