Merging Enterprise Manager Server (EMS)  prototype into trunk.  To start the server, launch the entry point com.l7tech.server.ems.Main.
The EMS requires Jetty 6.1.  As Jetty 6 is a significant upgrade from 5, with a totally new API, this had some impact on the SSB.  Hopefully the only externally-visible effect is to make the SSB faster.  

For Jetty 6 the proprietary API has been merged with the servlet API, so we could now provide full HttpServletRequestKnob and HttpServletResponseKnob services to client assertions just by inserting two lines in RequestHandler.  I haven't done this yet since I waht to review the existing client assertions first.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@18943 219f1c6e-4a0e-43c8-96f2-6e749be13fd4