updated sealsys with two new abilties.
update) updates itself from svn (actually via the fisheye URL) so we have a way of getting the latest version

customize) checks the system for the presence of the SCA card and/or the presence of the SCA software. If the card is not present, but the software is, then the startup script links are removed and the directory whose presence enables the configuration of an HSM via the configurator is renamed.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@19148 219f1c6e-4a0e-43c8-96f2-6e749be13fd4