bug 5501: The Gateway, Bridge, and Manager now properly handle certificates with multiple CN attributes in the subject or issuer DN.
The hostname verifier for outbound SSL will trust a host if any of the CNs in the cert match the expected hostname.

For UI purposes when displaying a subject or issuer name in a table cell, only a single CN value will be displayed.

When the XVC checks a client cert to see what the username is, it will avoid locking in a username if the client cert has multiple CNs.



git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@19958 219f1c6e-4a0e-43c8-96f2-6e749be13fd4