bug 5023: regex of variable now works per the func spec.  Regex assertions created by previous Gateway versions will continue to auto-target the request or the response depending on whether routing has happened, but all assertions created (or edited) from now on will use explicit targeting.
PolicyEnforcementContext.getTargetMessage() can now optionally work with variables that aren't actually of Message type.  If enabled, it will create a Message initialized from the variable and with a special StashManager that ensures changes made to the message get written back into the variable's value.

Introduced PolicyUtil to hold utility routines for working with policy trees.    Currently it provides a way to inspect an assertion, or a location in a policy, and see if it is pre or post routing.  (Code moved from Steve's routine in AddContextSensitiveAssertionAdvice.)

Server assertions can now throw an (unchecked) AssertionStatusException as an alternative to returning AssertionStatus.FAILED (or any other status).  This can make some server assertion code cleaner, but should not be used for failures that are expected to be frequent or routine.



git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@20713 219f1c6e-4a0e-43c8-96f2-6e749be13fd4