bug 6124: Turned off persistent session support in our embedded Tomcat since we don't use it.  It will now use /ssg/var/work as its work directory.  Tomcat will try to create this directory on startup if it doesn't already exist.  However, as we do not use JSP and  have now disabled persistent sessions, the directory will never have anything in it.
Added an alternate Gateway entry point GatewayForegroundMain that can be used for debugging clean Gateway shutdown when running from IDEA.  This starts a Gateway in the foreground with a tiny CLI that waits for someone to enter the "stop" command, then triggers a clean shutdown.

Fixed a number of background threads that were not being cleaned up properly to either shut down, or (in the case of the KerberosTicketRepository) at least to be marked as daemon threads so they won't interfere with shutdown.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@21528 219f1c6e-4a0e-43c8-96f2-6e749be13fd4