bug 6147: The SSM UI now does a case-insenstive match on the alias when checking whether to display the "default SSL" icon for a certificate.
To fix Bug #6039 we switched the Software DB keystore to use the Sun PKCS#12 implementation instead of the Bouncy Castle PKCS#12 implementation.  It turns out the Sun implementation converts aliases to lowercase, so the default SSL key's alias of "SSL" was not being seen by the SSM UI as matching the "ssl" in the list, so it was not displaying the [S] icon.  This was a cosmetic problem only and would clear up next time the default SSL alias cluster property was rewritten.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@21624 219f1c6e-4a0e-43c8-96f2-6e749be13fd4