bug 6820: There is now a friendlier error message when you attempt to export a key from an SCA 6000 HSM.
For now, I just added an isKeyExportSupported() method to SsgKeyFinder and had the HSM impl return false.  In the future we should remove this method and modify the key export so it won't require BouncyCastle to be registered as a security provider, and will be capable of succeeding any time the underlying private key implementation offered the capability of being wrapped for export.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@22826 219f1c6e-4a0e-43c8-96f2-6e749be13fd4