bug 6882: The message processing servlet no longer calls PolicyEnforcementContext.setVariable() on every request to set the request.compression.gzip.found.  Instead, this is now present as a built-in variable.  There is no user-visible effect of this change, but performance should increase by on the order of about half a percent.
In other news, we should make a note somewhere that PolicyEnforcementContext.setVariable() is far more expensive than one might tend to assume.  It does considerably more than just add a value to a HashMap due to our need to recognize built-in variables with various combinations of case and suffix insensitivity.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@23035 219f1c6e-4a0e-43c8-96f2-6e749be13fd4