bug 5502: Switch to a fixed-size thread pool to work around an apparent Tomcat problem.  Now, the minimum pool size is set to the same as io.httpMaxConcurrency, so if you set the max concurrency to 256. The Gateway still won't spawn all 256 threads on startup, however, as by default not even core threads are pre-started.
Our version of Tomcat uses a ThreadPoolExecutor with an unbounded work queue.  In that configuration the maximum pool size is effectively ignored, since ThreadPoolExecutor will only create new threads beyond the core size if the work queue is full.  Tomcat attempts to work around this by overridding LinkedBlockingQueue.offer() and doing some pool size management there, but the logic looks questionable and the end result is that the pool size was not grown when it should have been.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@23136 219f1c6e-4a0e-43c8-96f2-6e749be13fd4