bug 7194: Fixed based on initial fix from tactical_xacml_5_0 branch. Introduced ParameterizedMapping interface, which defines the find contract for any find logic which operates on TypeMapping s which implement it. Any TypeMapping implementation which accepts parameterized types e.g. List<String> should implement this, if it is envisioned that an instance of the mapping could be created with type parameters, which should be considered when finding TypeMapping s during the freeze/thaw process e.g. List<String> versus List<Integer>
Refactored TypeMapping Find logic into TypeMappingUtils.findTypeMapping(), which always check for ParameterizedType first when looking for TypeMappings. If a type is a ParameterizedType and ParameterizedMapping is not implemented by the TypeMapping, then the getRawType() method is called on the ParameterizedType and cast to a Class, so that the normal Class only finding mechanism can be used. If ParameterizedMapping is implemented, then ParameterizedMapping.getValueTypeClass() is called to get it's parameterized types, this is then used as part of the find logic to determine that the TypeMapping being examined is the correct TypeMapping for the Type being freezed/thawed.

Bug 7194 was reported as the first TypeMapping which could type map a List was found and used in place of the intended TypeMapping, which could type map a List<AttributeList>. The original CollectionTypeMapping class only contracted to use the valueType constructor parameter when deserializing as a type check, and not as part of the find process.

Changes work as when a Method is retrieved from a Class reflectively, the getGenericXXX() methods maintain the actual generic type information. The Type instances returned by these methods may be instances of ParameterizedType, which are not instances of Class, as Class extends Type and ParameterizedType is a sub interface of Type. The code: "Class typeClass = (object.type instanceof Class) ? (Class)object.type : (Class)((ParameterizedType)object.type).getRawType();" is used anywhere we need the Class of the Type parameter, and is refactored (from initial impl) into TypeMappingUtils.getClassForType().

Find logic in TypeMappingUtils.findTypeMapping() does not assume that a TypeMapping implementing ParameterizedMapping's getValueTypeClass() method only returns a single class. As a result this find logic will work with future TypeMappings which allow for a mapped object to be parameterized with more than 1 class e.g. a Map<String, Integer> versus Map<String, String>. 

Added test coverage for new TypeMappingUtils methods. Updated BeanTypeMapping.emitBeanProperties() to work with generic return / parameter types. Completed missing generic info on collections used by the RequestSwAAssertion assertion. 


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@23632 219f1c6e-4a0e-43c8-96f2-6e749be13fd4