Added the necessary WSDLs & XSDs to support the UDDIv3 subscription API.
Modified the 'generate-uddi-jar' target (used to build 'layer7-uddi-lib-1.0.1.jar') to include the subscription API classes.  Note that this change is commented out so that if the JAR is re-generated for some reason in Buzzcut these changes will not be included.

So, these changes lay the groundwork for enhancing our UDDI support, but no code changes were made.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@23696 219f1c6e-4a0e-43c8-96f2-6e749be13fd4