bug 7287: Fixed the issue - when an XPath expression evaluates to nothing or an error, the policy consumption is still not falsified.
In this bug fixing:
(1) Added a new checkbox in XPath Multiple Attribute panel.  The text of the checkbox is "Falsify policy if evaluating any required fields results in Not Found".

(2) If any required fields (ID, Data Type, and Issue Instant) evaluate as Not Found, then there are two results depending on the option "Falsify Policy":
- If the option is on, falsify policy.
- If the option is off, there are no <Attribute> element created in Subject, Resource, Action, or Environment and the iteration will continue.

Note: Issue Instant will be only considered in pre XACML 2.0 if it is specified.

(3) Added logging/auditing when an XPath expression evaluates to nothing (or says evaluating any required fields results in Not Found.)

If the option is off, the logging/auditing message is "A value for {0} was not found. Cannot add <Attribute> element to the {1} element", where {0} is one of ID, Data Type, and Issue Instant, and {1} is one of Subject, Resource, Action, and Environment.  The message LEVEL is INFO.

If the option is on, the logging/auditing message is "Policy falsified: a value for {0} was not found", where {0} is one of ID, Data Type, and Issue Instant.  The message LEVEL is INFO.  

(4) Modified the field 'Value' to allow it evaluated as empty (i.e., not found), so in the generated request, an Attribute element will still be created with an empty AttributeValue element.

(5) Fixed an issue - the value of Issue Instant cannot be loaded and saved in XACML 1.1.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24023 219f1c6e-4a0e-43c8-96f2-6e749be13fd4