bug 7544: Added the ability to see additional elliptic cuve names in the New Private Key dialog if the SSM is running with the undocumented system property "com.l7tech.console.keyStore.showAdditionalNamedCurves" set to "true".
This behavior is not sufficient to fix Bug #7544 and is disabled by default.

(Among the issues with enabling this system property: There is no indication given in the UI that (for example) secp256r1, P-256, and prime256v1 all refer to the same curve, so a user may be surprised if they create a "P-256" curve and see the resulting cert shows "secp256r1" instead.  The UI does not know about, and can't help you with, a provider such as Luna that recognizes some curve name synonyms but not others -- a user will have to try all the synonyms to find the one that is recognized.  The list is long and disorganized.  The long list duplicates secp384r1.)

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24147 219f1c6e-4a0e-43c8-96f2-6e749be13fd4