bug 7508: Finishing off non-SOAP XML security for the time being.  It still needs additional GUI work (form validation, dialog for VariableCredentialSourceAssertion), a bunch of policy validators, a bunch more unit tests, and a func spec or two, but I'm going to stop here for now.
Introduced lightweight DOM-only assertion Select Single Element by XPath.   This is easier to use to select a single Element variable than the multiple elements functionality added to the regular XPath assertions for Bug #4298.  However due to this overlap we may choose not to retain the SelectElementAssertion. 

Introduced new assertions for checking the output of a non-SOAP decrypt or verify:  Index Lookup by Item, and Item Lookup by Index.   The former is basically a grep of a multivalued context variable that return matching index numbers.  The latter is basically an array dereference that extracts a single value from a multivalued context variable.  (This could already be done with ExpandVariables syntax, but only for strings and not certificates etc., and seemingly not with an index that comes from a variable.)

Introduced  a VariableCredentialSourceAssertion that can be used to gather X.509 credentials from a certificate variable from an arbitrary source to use for authentication with a subsequent SpecficUser or MemberOfGroup assertion.  This can be used to authenticate users by their non-SOAP XML signatures.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24171 219f1c6e-4a0e-43c8-96f2-6e749be13fd4