bug 7577:  The WSDL proxy no longer pays attention to disabled credential source assertions (or composite assertions).
The WspReader now requires users to specify whether they want disabled assertions to be included in the parsed policy.  This should make it much easier to avoid problems in the future due to the presence of disabled assertions in contexts where there is no reason for them to be present.



git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24214 219f1c6e-4a0e-43c8-96f2-6e749be13fd4