bug 7646:  The CertValidationProcessorImpl no longer throws an IllegalStateException if there is no default policy designated after processing an EntityInvalidationEvent.
When the default is being changed, two updates are done: one update to clear the default flag from the old default entity, and one update to add the default flag to the new entity.  Depending on the ordering of the two EntityInvalidationEvents this can result in a brief window where there is no default policy.  Thus, this is not really an illegal state, and so should not result in an illegal state exception.

When there is no default policy designated the system uses a fallback policy which always fails.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24236 219f1c6e-4a0e-43c8-96f2-6e749be13fd4