bug 7367: Merging Franco's fix to trunk.  The cluster property has been renamed to "io.httpsAcceptedClientCa" but in all other ways the behavior is substantially as it was in Franco's patch.
This cluster property is hidden in the SSM UI.  However, if you manually type it in, the SSM will display the description, confirming that it recognizes the cluster property and showing you how to use it.

Note that if there is more than one PEM certificate the values must be comma delimited.  This differs from the usual convention for storing a bunch of PEM certs in a single text file (letting the PEM markers serve as their own delimiters).

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24322 219f1c6e-4a0e-43c8-96f2-6e749be13fd4