Renaming for the SAML Assertion.
Added support for existing assertions to use meta data for the context click actions and property dialog titles. Added class LegacyAssertionPropertyDialog which RequireWssSamlPropertiesPanel extends. This new class is a simple way for dialogs to inherit the use of their assertions meta data for setting up the title of their dialog. New assertions should not use this class but should instead work with AssertionPropertiesOkCancelSupport or a subclass. 

Updated NodeAction to override getName() from SecureAction with a default implementation which uses meta data for it's value. Added default getter for PROPERTIES_ACTION_NAME, which falls back on SHORT_NAME. Updated NodeAction and SecureAction to provide constructors which allow BaseAction's single argument constructor to be called, which does not call setActionValues(). This allows subclasses of NodeAction to specify their preference for setActionValues(). If the subclass wants to utilize meta data and use the inherited getName() from NodeAction, then they should flag this in the call to super(), which stops setActionValues() from being called before instance variables are set


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24396 219f1c6e-4a0e-43c8-96f2-6e749be13fd4