bug 7590: Merging Steve's patch to trunk after testing on Windows and with the SSM applet:
This patch makes modular assertion class loading more efficient and smarter. 

When resources are requested the server can choose to respond with a set of
related resources instead of a single resource. For modular assertions the set
is a single package (not including sub-packages).

This means there are fewer round trips and the client knows that it has all the
resources related to 'X' package so doesn't request resources that don't exist
from the server (such as resource bundles, or tree nodes)

For me, this doesn't make much difference to the log in speed, but opening
policies or properties dialogs is faster (since the classes are already
loaded).


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24397 219f1c6e-4a0e-43c8-96f2-6e749be13fd4