bug 5752: Added support for overridding the request HTTP method to the HTTP Routing Assertion properties dialog.  There are a pair of radio buttons (for use default vs. override) and a combobox for selecting GET, POST, PUT, or DELETE.  (HEAD is not listed because it is not supported by the Commons HTTP client and hence would always fail at runtime if selected).
The combobox is disabled when override is not selected.  The override radio button is disabled for the Bridge Routing Assertion, which does not support the feature.

When automatic selection of the request method is selected the server assertion behaves as it did pre-5.2: using the method from the request, if any, and defaulting to POST if not set (as in the case of a request that arrived over a non-HTTP transport, or a Message context variable).


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24573 219f1c6e-4a0e-43c8-96f2-6e749be13fd4