bug 7544: Additional EC named curves are now offered in the SSM's "New Private Key" dialog.  Not all curves will work with all keystors and crypto providers.  Not all curves will work with all hash algorithms.  (On the Luna you'll get an error if you try to use a signature hash whose output is larger than the curve size, ie trying to use SHA-384 with a sect163k1 curve.)
This changeset also addresses Bug #7865: SignerInfo now has custom writeObject() and readObject() methods.  The readObject() method will attempt to recover and continue anyway if a certificate cannot be parsed.  It will replace the certificate with null.  Added a parallel array of subject DN strings, so the client will at least know the DN of the certificate it can't see.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24617 219f1c6e-4a0e-43c8-96f2-6e749be13fd4