bug 7623: The Gateway now loads either the FIPS or non-FIPS version of RSA BSAFE Crypto-J dynamically at runtime, depending on whether FIPS mode is enabled on not.
Using RSA software crypto in FIPS mode, use of an ECC client cert for outbound TLS from the Gateway will continue to fail with "error signing certificate verify".

In non-FIPS mode, this will now work as expected.  (It continues to work as it did before when using a Luna HSM as the crypto provider.)

Note that performance of the new 4.1.0.1 version of the RSA library is actually slightly lower under high concurrency than it was with 4.0, as RSA appear to have added some additional locking (perhaps to fix other bugs?).  This means the Gateway might actually be slightly faster in FIPS mode than in non-FIPS mode under highly concurrent crypto workloads.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24701 219f1c6e-4a0e-43c8-96f2-6e749be13fd4