bug 8027: fixed by correcting logic check when publishing a Gateway WSDL to UDDI.
All publishing tasks now catch any Exceptions and if applicable rethrows them as an UDDIHandledTaskException. Calling code can then handle this exception and call it's handleError method in a new transaction so that the publish status can always be updated regardless of the outcome of the publish task.
ServiceUDDISettings dialog changes: Added in warning message which stops the user from clicking on 'Clear' when a proxy bindingTemplate or the service has been overwritten. Added logic to disable the 'Select' button when a related business service has been chosen. Updated viewToModel to not dismiss the dialog if the user chosen to select 'No' on a confirmation dialog. 


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24945 219f1c6e-4a0e-43c8-96f2-6e749be13fd4